requirements.txt
appdirs==1.4.4
appnope==0.1.2
audioread==2.1.9
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.0
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.35.0
matplotlib==3.3.3
mypy-extensions==0.4.3
numba==0.52.0
numpy==1.19.4
packaging==20.8
parso==0.7.1
pathspec==0.8.1
peewee==3.14.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pooch==1.3.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pychorus==0.1
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
youtube-dl==2020.12.22
